exo1-exo3

+++ ex00_json2ini/jsonConversion.py
@@ -1,14 +1,16 @@
 import sys, json, os.path
 
-output = ""
-with open(sys.argv[1]) as json_file:
-    data = json.load(json_file)
-    for key in data.keys():
-        output += "\n"+"["+key+"]\n"
-        for sub_key,value in data[key].items():
-            output += sub_key + "=" + value + "\n"
+if(os.path.exists(sys.argv[1])):
+    output = ""
+    with open(sys.argv[1]) as json_file:
+        data = json.load(json_file)
+        for key in data.keys():
+            output += f"\n[{key}]\n"
+            for sub_key,value in data[key].items():
+                output += f"{sub_key}={value}\n"
 
-    print(output)
-    
-with open("output.ini", "w") as outputFile:
-	outputFile.write(output)
+        print(output)
+
+    with open("output.ini", "w") as outputFile:
+        outputFile.write(output)
+        print("Job Done")
